Move filemap to crane
