+++ pkg/crane/filemap.go
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package filemap allows you to create images directly from filemaps.
-package filemap
+package crane
 
 import (
 	"archive/tar"
